# swagger.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  contact: {}
  version: '1.0'
servers:
- url: https://e-commerce-api-tk1s.onrender.com/api/v1
  variables: {}
paths:
...
  /product:
    get:
      tags:
      - Products
      summary: Get All Products
      operationId: GetAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Products
      summary: Create Product
      operationId: CreateProduct
      parameters: []
      requestBody:
        description: '"name", "price", "boughtYear" and "brand" are necessary to provide.<br>"description" is optional.<br>You must be logged in to create product'
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateProductRequest'
              - example:
                  name: albany sectional
                  price: 110
                  boughtYear: 2005
                  brand: liddy
                  description: Hello, Wolrd!
            example:
              name: albany sectional
              price: 110
              boughtYear: 2005
              brand: liddy
              description: Hello, Wolrd!
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /product/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: Specify product id
    get:
      tags:
      - Products
      summary: Get Single Product
      operationId: GetSingleProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    delete:
      tags:
      - Products
      summary: Delete Product
      operationId: DeleteProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
      - Products
      summary: Update Product
      operationId: UpdateProduct
      parameters: []
      requestBody:
        description: 'Only "name", "brand", "price", "boughtYear" and "description" can be updated.<br>A product can be updated or deleted only by its owner'
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateProductRequest'
              - example:
                  boughtYear: 2018
            example:
              boughtYear: 2018
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateProductRequest:
      title: CreateProductRequest
      required:
      - name
      - price
      - boughtYear
      - brand
      - description
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        boughtYear:
          type: integer
          format: int32
        brand:
          type: string
        description:
          type: string
      example:
        name: albany sectional
        price: 110
        boughtYear: 2005
        brand: liddy
        description: Hello, Wolrd!
    UpdateProductRequest:
      title: UpdateProductRequest
      required:
      - boughtYear
      type: object
      properties:
        boughtYear:
          type: integer
          format: int32
      example:
        boughtYear: 2018
...